+++ pages/page-privacy-policy.php
@@ -1,0 +1,9 @@
+<article class="privacy-policy">
+<div class="post-content">
+<?php
+while ( have_posts() ) : the_post();
+	the_content();
+endwhile;
+?>
+</div>
+</article>