+++ parts/footer.php
@@ -1,7 +1,7 @@
 	<footer>
 		<div class="wp-group">
 			<p><?php echo date('Y'); ?>, Cigno Zen</p>
-			<p><a href="<?php echo esc_url(home_url('/contatti')); ?>">Contatti</a></p>
+			<p><a href="<?php echo esc_url(home_url('/contatti')); ?>">Contatti</a> | <a href="<?php echo esc_url(home_url('/privacy-policy')); ?>">Privacy</a></p>
 		</div>
 	</footer>
 	<?php wp_footer(); ?>